Fix:다중잡 설정
