+++ src/main/java/TMT/Ranking/TmtRankServerApplication.java
@@ -1,5 +1,6 @@
 package TMT.Ranking;
 
+import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
