+++ src/main/java/TMT/Ranking/batch/presentation/DailyRankingBatchScheduler.java
@@ -1,6 +1,8 @@
 package TMT.Ranking.batch.presentation;
 
 import TMT.Ranking.assetranking.presentation.AssetRankingBatchJobConfig;
+import TMT.Ranking.monthlyranking.presentation.MonthlyRankingJobConfig;
+import TMT.Ranking.weeklyranking.presentation.WeeklyRankingJobConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.Job;
@@ -21,11 +23,14 @@ public class DailyRankingBatchScheduler {
     private final JobRepository jobRepository;
     private final PlatformTransactionManager platformTransactionManager;
     private final AssetRankingBatchJobConfig assetRankingBatchJobConfig;
+    private final WeeklyRankingJobConfig weeklyRankingJobConfig;
+    private final MonthlyRankingJobConfig monthlyRankingBatchConfig;
 
-    @Scheduled(cron = "0 39 16 ? * MON-FRI") //일간 수익률 집계
+    @Scheduled(cron = "0 36 19 ? * MON-FRI") //일간 수익률 집계
     public void dailyRankingBatchStart()
             throws Exception{
 
+        log.info("start dailyRankingBatch");
         jobLauncher.run((Job) dailyRankingJobConfig.dailyRankingJob(
                 jobRepository,platformTransactionManager), new JobParametersBuilder()
                 .addLong("time", System.currentTimeMillis())
@@ -37,7 +42,7 @@ public class DailyRankingBatchScheduler {
     public void weeklyRankingBatchStart()
             throws Exception{
 
-        jobLauncher.run((Job) dailyRankingJobConfig.weeklyRanking(
+        jobLauncher.run((Job) weeklyRankingJobConfig.weeklyRanking(
                 jobRepository,platformTransactionManager), new JobParametersBuilder()
                 .addLong("time", System.currentTimeMillis())
                 .toJobParameters());
@@ -48,7 +53,7 @@ public class DailyRankingBatchScheduler {
     public void monthlyRankingBatchStart()
             throws Exception{
 
-        jobLauncher.run((Job) dailyRankingJobConfig.monthlyRanking(
+        jobLauncher.run((Job) monthlyRankingBatchConfig.monthlyRanking(
                 jobRepository,platformTransactionManager), new JobParametersBuilder()
                 .addLong("time", System.currentTimeMillis())
                 .toJobParameters());
